Create Pointers
Write a short C program that declares and initializes (to any value you
like) a double, an int, and a char. Next declare and initialize a pointer
to each of the three variables. Your program should then print the address of, and value stored in, and the memory size (in bytes) of each of
the six variables. Use the “0x%x” formatting specifier to print addresses
in hexadecimal. You should see addresses that look something like this:
”0xbfe55918”. The initial characters ”0x” tell you that hexadecimal notation is being used; the remainder of the digits give the address itself. Use
“%f” to print a floating value. Use the sizeof operator to determine the
memory size allocated for each variable.